chore: split mega test file into three parts
